requirements.txt
Flask==0.12.3
click==6.7
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pytz==2017.2
requests<3.0.0
Werkzeug==0.12.1
google-cloud-storage==1.13.2
nltk>=3.0.0
wtforms>=2.0.0
scikit-learn>=0.18